:art: Moves childAtPosition to CustomMatcher class
